+++ app.py
@@ -1,7 +1,16 @@
 from flask import Flask, render_template
-
+from flask_sqlalchemy import SQLAlchemy
+from forms import LoginForm, RegisterForm
 
 app = Flask(__name__)
+
+
+
+
+db = SQLAlchemy(app)
+
+
+
 
 @app.route("/")
 @app.route("/home")
@@ -12,7 +21,9 @@ def index():
 def login():
     return render_template("login.html")
 
-
+@app.route("/register")
+def register():
+    return render_template("register.html")
 
 
 if __name__ == "__main__":